Move GoalTypeOption enum to separate class.
