+++ src/main/java/swaps/view/SwapsPresenter.java
@@ -1,0 +1,11 @@
+package swaps.view;
+
+public class SwapsPresenter {
+    private SwapsView view;
+
+    public SwapsPresenter() {
+        view = new SwapsView();
+    }
+
+
+}
